4 feb 2018 1. #111 - user can see whom he follows and who is following him in the People section of navigation drawer. 2. Added date time in User joined timeline.

+++ app/src/main/java/com/cookery/utils/InternetUtility.java
@@ -710,7 +710,7 @@ public class InternetUtility {
             Date end = new Date();
 
             Log.i(CLASS_NAME, "*");
-            Log.i(CLASS_NAME, "*** POST ("+(end.getTime() - start.getTime())/1000+" seconds)***");
+            Log.i(CLASS_NAME, "*** POST ("+(double)(end.getTime() - start.getTime())/1000+" seconds)***");
             Log.i(CLASS_NAME, "URL : "+url);
             Log.i(CLASS_NAME, "PARAMS : "+paramMap);
             Log.i(CLASS_NAME, "RESPONSE : "+response);
